requirements.txt
gunicorn
certifi
discord.py
finance_datareader
Flask
Flask_Cors
Flask_Discord
GitPython
investpy
matplotlib
mojito2
mplchart
nest_asyncio
numpy==1.26.4
pandas==2.2.1
pandas_ta==0.3.14b0
Pillow
python-dotenv
PyYAML
quart
quart_cors
Requests
tabulate
tradingview_ta
yfinance
plotly
google-cloud-secret-manager
seaborn
openai
google-generativeai
pandas_market_calendars
scikit-learn
squarify
pywin32==308